% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorResponse.R
\name{resampleBands}
\alias{resampleBands}
\title{resampleBands}
\usage{
resampleBands(wl, spectra, sensor, ...)
}
\arguments{
\item{wl}{a vector of wavelengths}

\item{spectra}{a vector of reflectance/radiance values}

\item{sensor}{sensor bands to simulate / resample}
}
\description{
Function for spectral response modelling of common
sensor platforms
}
