% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorResponse.R
\docType{data}
\name{response_functions}
\alias{response_functions}
\title{Data on sensor response function for various Remote Sensing Platforms}
\usage{
data(response_functions)
}
\description{
Data on sensor response function for various Remote Sensing Platforms
}
\section{details}{

***********************************************************************
response_functions (Jan, 20th 2020)
 The dataset contains the following response functions (in a list):
***********************************************************************
 \itemize{

\item [1] = Hyperion, data gives mean (nw) and sigma of best fitting Gaussuin
for each band (note that band differ for each of the 256 spatial pixels!)
This data uses averages. Data was originally supplied by the EO-1 Instrument Teams.
http://www.eoc.csiro.au/hswww/oz_pi/specresp.htm
}
}

\section{references}{

}

\keyword{datasets}
